% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.fmtsas.R
\name{extract.fmtsas_c}
\alias{extract.fmtsas_c}
\alias{[.fmtsas_c}
\title{Opérateur \verb{[]} pour objet fmtsas_c}
\usage{
\method{[}{fmtsas_c}(x, i, keep_na = FALSE)
}
\arguments{
\item{x}{objet sur lequel on veut procéder à une extraction/conversion.}

\item{i}{noms des éléments à extraire ou convertir.}

\item{keep_na}{conserver les valeurs manquantes telles quelles. Par défaut
\code{FALSE} : les \code{NA} seront remplacés par la valeur \code{other} (comme dans SAS).}
}
\value{
Un vecteur caractère non nommé, de même taille que le vecteur passé
en paramètre \code{i}.
}
\description{
Variante de \verb{[]} pour objet fmtsas_c, prenant en compte une possible valeur
par défaut.
}
\details{
L'opérateur \verb{[]} pour un objet "fmtsas_c" s'utilise de la même façon que le
\verb{[]} habituel. La différence est qu'il remplace les valeurs inconnues par la
valeur définie dans l'attribut \code{"other"}.

Autres particularités :
\itemize{
\item il ne s'utilise qu'avec un vecteur caractère en paramètre (pas d'entiers,
ni de booléens). Un facteur est aussi possible, il sera transformé en
caractère ;
\item le résultat est dépourvu de noms d'éléments.
}
}
\examples{
# construit un objet fmtsas_c
conv <- fmtsas_c(c("A" = "LIB_A", "B" = "LIB_B"), other = "?")

# effectue une conversion
conv[c("A", "A", "B", "C", NA)]
conv[c("A", "A", "B", "C", NA), keep_na = TRUE]
}
