% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtsas_class.R
\name{fmtsas_c}
\alias{fmtsas_c}
\title{Constructeur d'objet S3 fmtsas_c}
\usage{
fmtsas_c(x, other = NA_character_)
}
\arguments{
\item{x}{vecteur caractère dont tous les éléments sont nommés représentant
un format SAS (association valeur entrée = valeur sortie).}

\item{other}{modalité par défaut. Un vecteur caractère de longueur 1.}
}
\value{
Un objet de classes \code{"fmtsas_c"} (et \code{"fmtsas"}). Ces objets seront
en pratique créés par \link{from_tab} ou \link{from_pgm} (ces fonctions génèrent
une liste d'objets de classe \code{fmtsas_c}).
}
\description{
Construit un objet représentant un format caractère SAS en R, à partir de
ses composantes (vecteur d'associations et modalité par défaut).
}
