% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_tab.R
\name{from_tab}
\alias{from_tab}
\title{Convertit une table pour format SAS}
\usage{
from_tab(sas_data)
}
\arguments{
\item{sas_data}{un data.frame importé depuis une table SAS pour proc format.
Ce data.frame doit contenir a minima les colonnes \code{FMTNAME}, \code{START},
\code{LABEL} et \code{TYPE} (majuscule ou minuscule).}
}
\value{
Une liste contenant autant d'éléments que de formats si les données
avaient été générées par une proc format.
\itemize{
\item les noms de la liste correspondent aux noms des formats (\code{FMTNAME}) ;
\item les éléments de la liste sont des vecteurs contenant les relations entre
valeurs initiales et valeurs converties.
}

Voir les exemples pour l'utilisation de cette liste.
}
\description{
Convertit les données utilisées pour une proc format (option CNTLIN=) en une
liste de vecteurs qui pourra servir à effectuer des conversions.
}
\details{
La table en entrée est un data.frame, à importer dans R à partir d'une table
SAS (par exemple avec le package \strong{haven}).

Cette fonction ne gère pour l'instant que des formats caractère. De plus,
s'il y a une colonne \code{END} dans la table en entrée, elle doit être égale à la
colonne \code{START}.
}
\examples{
# donnees en entree (provenant d'une table SAS)
format_data <-
  data.frame(
    FMTNAME = c("v1fmt", "v1fmt", "v1fmt", "v2fmt", "v2fmt"),
    TYPE    = c(    "C",     "C",     "C",     "C",     "C"),
    START   = c(    "1",     "2",     "3",     "A",     "B"),
    END     = c(    "1",     "2",     "3",     "A",     "B"),
    LABEL   = c("Lib 1", "Lib23", "Lib23", "Lib A", "Lib B"),
    stringsAsFactors = FALSE
  )

conv <- from_tab(format_data)

# Utilisation :

# soit un jeu de donnees contenant des codes a convertir en libelles
donnees <-
  data.frame(
    VAR1_CODE = c("1", "2", "3", "1", "Z"),
    stringsAsFactors = FALSE
  )

# pour remplacer les codes par les libelles (pour VAR1)
donnees$VAR1_LIB <- conv$v1fmt[donnees$VAR1_CODE]

donnees
}
