% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmtsas_class.R
\name{fmtsas_c}
\alias{fmtsas_c}
\title{Objet de classe S3 "fmtsas_c"}
\usage{
fmtsas_c(x, other = NULL)
}
\arguments{
\item{x}{vecteur caractère (dont tous les éléments sont nommés) représentant
un format SAS (association \emph{valeur(s) entrée = valeur sortie}).}

\item{other}{modalité par défaut (paramètre optionnel). Un vecteur caractère
de longueur 1.}
}
\value{
Un objet de classe \code{"fmtsas_c"} (et donc \code{"fmtsas"}), c'est-à-dire un
vecteur nommé possédant un éventuel attribut "other".
}
\description{
Construit un objet représentant en R un format caractère SAS, à partir de
ses composantes (vecteur d'associations et modalité par défaut).
}
\details{
La classe S3 \code{"fmtsas_c"} représente un format caractère SAS. Cette classe
hérite de la classe \code{"fmtsas"}, qui représente tout type de format
(caractère, numérique...).

Un objet \code{"fmtsas_c"} est un vecteur nommé contenant les relations entre
valeurs initiales et valeurs converties. Il possède en outre un éventuel
attribut "other", où est stockée une valeur par défaut. Cette valeur sera
utiliséelors d'une conversion avec
l'\link[=extract.fmtsas_c]{opérateur de sélection}(\verb{[]}).

Le nom du format n'est pas stocké dans ce type d'objet. Il pourra être
spécifié comme un des noms d'une liste de \code{fmtsas}. Une telle liste sera en
pratique générée par \link{from_tab} ou \link{from_pgm}.
}
\examples{
fmtsas_c(c("A" = "LIBA", "B" = "LIB_B"))
fmtsas_c(c("A" = "LIBA", "B" = "LIB_B"), other = "??")
}
